fix(api): split system memory monitoring route
