+++ crates/routa-server/src/api/mod.rs
@@ -112,7 +112,8 @@ pub fn api_router(state: AppState) -> Router<AppState> {
         .nest("/api/sandboxes", sandbox::router())
         .nest("/api/specialists", specialists::router())
         .nest("/api/spec", spec::router())
-        .nest("/api/memory", memory::router())
+        .nest("/api/system/memory", memory::router())
+        .nest("/api/memory", memory::legacy_router())
         .nest("/api/debug", debug::router())
         .nest("/api/polling", polling::router())
         .nest("/api/workflows", workflows::router())
